+++ src/js/filters/PresenterFilters.tsx
@@ -1,50 +1,19 @@
 import { useCallback, useContext, useMemo } from 'react';
-import Select, { MultiValue, Theme, StylesConfig } from 'react-select';
+import { MultiValue } from 'react-select';
 import { Option, Presenter } from '../types';
 import { useDb } from '../dbHooks';
 import { formatName } from '../utils';
-import { Colors, hosts } from '../constants';
+import { hosts } from '../constants';
 import { FilterSection } from './FilterSection';
 import { FiltersContext } from './FiltersContext';
+import { DropdownMultiselect } from './DropdownMultiselect';
+import { sortByLabel } from '../utils';
 
 const noOneFound = () => 'No one found';
 
 type PresenterType = 'host' | 'qiElf' | 'guest';
 
 type GroupedPresenterOptions = Record<PresenterType, Presenter[]>;
-
-const sortByLabel = (a: Option, b: Option) =>
-  a.label.toLocaleLowerCase().localeCompare(b.label.toLowerCase());
-
-const customTheme = (theme: Theme) => ({
-  ...theme,
-  colors: {
-    ...theme.colors,
-    primary: Colors.dimGrey,
-    primary25: Colors.lightBlue,
-    neutral0: Colors.citrineLighter,
-    neutral10: Colors.lightBlue,
-    neutral20: Colors.night,
-    neutral30: Colors.night,
-    neutral40: Colors.dimGrey,
-    neutral50: Colors.dimGrey,
-    neutral60: Colors.night,
-    danger: Colors.slateGrey,
-    dangerLight: Colors.lighterBlue,
-  },
-});
-
-const styles: StylesConfig<Option, true> = {
-  container: base => ({
-    ...base,
-    flexGrow: 1,
-  }),
-  menu: base => ({
-    ...base,
-    textAlign: 'left',
-    whiteSpace: 'nowrap',
-  }),
-};
 
 const groupLabels: Record<PresenterType, string> = {
   host: 'Hosts',
@@ -106,15 +75,10 @@ export const PresenterFilters = () => {
   return (
     presenters && (
       <FilterSection label="Presenter filters:">
-        <Select
+        <DropdownMultiselect
           defaultValue={defaultValue}
-          theme={customTheme}
-          styles={styles}
           noOptionsMessage={noOneFound}
           placeholder="Search presenters"
-          isSearchable={true}
-          isClearable={true}
-          isMulti={true}
           options={options}
           onChange={handleChange}
         />
