--- src/js/filters/filterUtils.ts
@@ -1,7 +1,0 @@
-import { Venue } from '../types';
-
-export const formatVenueName = (venue: Venue) => {
-  return `${venue.name} (${[venue.region, venue.city, venue.state]
-    .filter(n => n)
-    .join(', ')})`;
-};
